rag_project/requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.3.0
tiktoken>=0.4.0
faiss-cpu>=1.7.4
pypdf>=3.15.1
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
langchain-community